converted all js files to jsx files. Updated webpack entry files. Renamed several files and moves styles to one .css file.
