+++ frontend/canvas-chrome-ext/package.json
@@ -17,6 +17,7 @@
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
     "babel-loader": "^9.1.2",
+    "copy-webpack-plugin": "^11.0.0",
     "html-webpack-plugin": "^5.5.0",
     "react-dom": "^18.2.0",
     "webpack": "^5.75.0",
