+++ frontend/canvas-chrome-ext/public/studentSubmission.json
@@ -1,0 +1,45 @@
+{
+  "submissionId": 2543245,
+  "studentId": 423423,
+  "assignmentId": 2342342,
+  "submissionFiles": [
+    {
+      "name": "main.cpp",
+      "fileContent": [
+        "#include <iostream>\n",
+        "\n",
+        "int main() {\n",
+        "    Hello hello;\n",
+        "    hello.sayHello();\n",
+        "    return 0;\n",
+        "}"
+      ]
+    },
+    {
+      "name": "Hello.h",
+      "fileContent": [
+        "#pragma once\n",
+        "\n",
+        "class Hello {\n",
+        "public:\n",
+        "    void sayHello();\n",
+        "}"
+      ]
+    },
+    {
+      "name": "Hello.cpp",
+      "fileContent": [
+        "#include <iostream>\n",
+        "#include \"Hello.h\"\n",
+        "\n",
+        "using namespace std;\n",
+        "\n",
+        "void Hello::sayHello() {\n",
+        "    cout << \"Hello World!\" << endl;\n",
+        "}"
+      ]
+    }
+  ],
+  "submissionDirectory": "129373263"
+}
+
