+++ frontend/canvas-chrome-ext/src/html/compilation-notification.html
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
-    <script src = "../scripts/notification.js"></script>
-    <link rel="stylesheet" href ="../styles/compilation_notification.css">
+    <script src = "../scripts/compilation-notification.jsx"></script>
+    <link rel="stylesheet" href ="../extension-styles.css">
 </head>
 <body>
 
@@ -12,7 +12,7 @@
       <h1 class ="results-display">Compilation Results</h1>
 
 
-      <div class = "results-container">
+      <div class = "compilation-notification-results-container">
           <p id ="update-text"> Results of Compilation </p>
       </div>
