+++ frontend/canvas-chrome-ext/src/html/instructor-view.html
@@ -5,8 +5,10 @@
     <title>HELLO WORLD</title>
     <p>Visible text</p>
     <p class="secret">Invisible text</p>
+    <link rel="stylesheet" href="../extension-styles.css">
 </head>
-<body>
 
+<body>
+    <div  id="react-target"></div>
 </body>
 </html>
