+++ frontend/canvas-chrome-ext/src/html/popup.html
@@ -5,9 +5,9 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Canvas Code Grader</title>
-  <link rel="stylesheet" href="../styles/popup.css">
+  <link rel="stylesheet" href="../extension-styles.css">
 
-  <script src = "../popup.js"> </script>
+  <script src = "../popup.jsx"> </script>
 </head>
 
 <body>
