--- frontend/canvas-chrome-ext/src/styles/InstructorView.css
@@ -1,3 +1,0 @@
-p.secret {
-    display: none;
-}
