--- frontend/canvas-chrome-ext/src/styles/submission_read_only_view.css
@@ -1,13 +1,0 @@
-.a {
-    display: contents;
-    border: 2px solid red;
-    background-color: #ccc;
-    padding: 10px;
-    width: 200px;
-}
-
-.b {
-    border: 2px solid blue;
-    background-color: lightblue;
-    padding: 10px;
-}
